# 🧹 Pre-initialize all compress handlers