--- a/middleware/compress_test.go
+++ b/middleware/compress_test.go
@@ -144,7 +144,7 @@ func Test_Middleware_Compress_Panic(t *testing.T) {
 	Compress("invalid")
 }
 
-// go test -v -run=^$ -bench=Benchmark_Middleware_Compress -benchmem -count=4
+// go test -v ./... -run=^$ -bench=Benchmark_Middleware_Compress -benchmem -count=4
 func Benchmark_Middleware_Compress(b *testing.B) {
 	app := fiber.New()
 	app.Use(Compress())